+++ test/worker/index.js
@@ -1,8 +1,9 @@
 const config = require('config')
-const connection = config.connection
+const connection = config.connection || {}
+connection.pkg = 'redis'
 const Worker = require('node-resque').worker
-const Redis = require('ioredis')
-const redis = new Redis(connection)
+const Redis = require('redis')
+const redis = Redis.createClient(connection)
 
 const jobs = {
   succeed: {
@@ -16,7 +17,7 @@ const jobs = {
 const queues = ['koop']
 const worker = new Worker({connection, queues}, jobs)
 worker.connect(() => worker.start())
-worker.on('end', () => redis.end())
+worker.on('end', () => redis.quit())
 
 function fail (job, done) {
   redis.publish('jobs', JSON.stringify({id: job.job_id, status: 'fail'}))