+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* Replace ioredis with node-redis
+
 ## [0.1.0] - 2016-03-16
 ### Added
 * Added scheduler for delayed jobs